LayoutTests/inspector/css/selector-specificity-expected.txt

Testing that selectors have expected specificity values.

* (0, 0, 0)
h1 (0, 0, 1)
.class (0, 1, 0)
#id (1, 0, 0)
body h1.class-one.class-two (0, 2, 2)
body #foo (1, 0, 1)
body > #foo (1, 0, 1)
body > #foo.a.b (1, 2, 1)
h1::before (0, 0, 2)
body h1::before (0, 0, 3)
body.a h1.b::before (0, 2, 3)
